filterbutton centered
i was having lunch and suddenly it hit me that i'm using margins inside a frame layout and i had a epiphany that what if i used relative layout and centered button inside the parent? so that was it. i did it. and it works.

yeah i mess things up :confused: